requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
idna==3.6
mysql-connector-python==8.3.0
mysql-connector-python-rf==2.2.2
numpy==1.26.4
opencv-python==4.9.0.80
pycparser==2.21
pyserial==3.5
requests==2.31.0
setuptools==69.1.0
sms==0.4
smsir==1.0.5
smsir-python==1.0.7
smsiran==0.0.5
urllib3==2.2.0
